requirements.txt
requests
edge-tts
openai-whisper
moviepy
pydub
